Create README.md
poen_stent_proe.prt is a Creo-Pro/Engineer part file of a parameterized version of the Open Stent Design project. Included is a PowerPoint presentation that describes in some detail the CAD modeling steps used. The file stent_config.pro is a Pro/Engineer configuration file to be used with the model. The file proe_open_stent_layer_file.pro is a pro/e layer configuration file that will help in visualizing the CAD model. The CAD model is designed for intermediate to advanced users of Creo-Pro/engineer.